+++ butterfly-utilities/src/main/java/com/paypal/butterfly/utilities/operations/text/InsertText.java
@@ -6,6 +6,8 @@ import com.paypal.butterfly.extensions.api.TransformationOperation;
 import com.paypal.butterfly.extensions.api.exception.TransformationDefinitionException;
 import com.paypal.butterfly.extensions.api.exception.TransformationOperationException;
 import com.paypal.butterfly.extensions.api.exception.TransformationUtilityException;
+import com.paypal.butterfly.utilities.operations.EolBufferedReader;
+import com.paypal.butterfly.utilities.operations.EolHelper;
 
 import java.io.*;
 import java.net.MalformedURLException;
@@ -13,8 +15,10 @@ import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.util.regex.Pattern;
 
+import static com.paypal.butterfly.utilities.operations.EolHelper.removeEol;
+
 /**
- * Operation to insert text into another text file.
+ * Operation to insert text from one file into another text file.
  * The text can be inserted:
  * <ol>
  *     <li>InsertionMode.CONCAT: At the final of the file (default)</li>
@@ -22,6 +26,10 @@ import java.util.regex.Pattern;
  *     <li>InsertionMode.REGEX_FIRST: Right after only the first line to match the specified regular expression</li>
  *     <li>InsertionMode.REGEX_ALL: Right after any line to match the specified regular expression</li>
  * </ol>
+ * Notice concat is the default insertion mode. It is also important to state that the text to be inserted will
+ * always start on a new line and, if not placed on the end of the file, the continuation of the original text
+ * will follow in a new line, even if the inserted text does not end with a line break.
+ *
  * @see {@link #setInsertionMode(InsertionMode)}
  * @see {@link InsertionMode}
  * @author facarvalho
@@ -160,39 +168,45 @@ public class InsertText extends TransformationOperation<InsertText> {
         return String.format(DESCRIPTION, textFileUrl.getFile(), getRelativePath());
     }
 
-    @Override
     @edu.umd.cs.findbugs.annotations.SuppressFBWarnings (value="NP_ALWAYS_NULL_EXCEPTION")
+    @Override
     protected TOExecutionResult execution(File transformedAppFolder, TransformationContext transformationContext) {
         File fileToBeChanged = getAbsoluteFile(transformedAppFolder, transformationContext);
+
+        if (!fileToBeChanged.exists()) {
+            // TODO Should this be done as pre-validation?
+            FileNotFoundException ex = new FileNotFoundException("File to be modified has not been found");
+            return TOExecutionResult.error(this, ex);
+        }
 
         File tempFile = new File(fileToBeChanged.getAbsolutePath() + "_temp_" + System.currentTimeMillis());
         BufferedReader readerOriginalFile = null;
         BufferedReader readerText = null;
         BufferedWriter writer = null;
-        String details;
         TOExecutionResult result = null;
 
         try {
+            final String eol = EolHelper.findEolDefaultToOs(fileToBeChanged);
+
             readerOriginalFile = new BufferedReader(new InputStreamReader(new FileInputStream(fileToBeChanged), StandardCharsets.UTF_8));
             readerText = new BufferedReader(new InputStreamReader(textFileUrl.openStream(), StandardCharsets.UTF_8));
             writer = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(tempFile), StandardCharsets.UTF_8));
 
             switch (insertionMode) {
                 case LINE_NUMBER:
-                    details = insertAtSpecificLine(readerText, readerOriginalFile, writer);
+                    result = insertAtSpecificLine(readerText, readerOriginalFile, writer, eol);
                     break;
                 case REGEX_FIRST:
-                    details = insertAfterRegex(readerText, readerOriginalFile, writer, true);
+                    result = insertAfterRegex(readerText, readerOriginalFile, writer, true, eol);
                     break;
                 case REGEX_ALL:
-                    details = insertAfterRegex(readerText, readerOriginalFile, writer, false);
+                    result = insertAfterRegex(readerText, readerOriginalFile, writer, false, eol);
                     break;
                 default:
                 case CONCAT:
-                    details = concat(readerText, readerOriginalFile, writer);
+                    result = concat(readerText, readerOriginalFile, writer, eol);
                     break;
             }
-            result = TOExecutionResult.success(this, details);
         } catch (IOException e) {
             result = TOExecutionResult.error(this, e);
         } finally {
@@ -220,12 +234,12 @@ public class InsertText extends TransformationOperation<InsertText> {
         boolean deleted = fileToBeChanged.delete();
         if(deleted) {
             if (!tempFile.renameTo(fileToBeChanged)) {
-                details = String.format("Error when renaming temporary file %s to %s", getRelativePath(transformedAppFolder, tempFile), getRelativePath(transformedAppFolder, fileToBeChanged));
+                String details = String.format("Error when renaming temporary file %s to %s", getRelativePath(transformedAppFolder, tempFile), getRelativePath(transformedAppFolder, fileToBeChanged));
                 TransformationOperationException e = new TransformationOperationException(details);
                 result = TOExecutionResult.error(this, e);
             }
         } else {
-            details = String.format("Error when deleting %s", getRelativePath(transformedAppFolder, fileToBeChanged));
+            String details = String.format("Error when deleting %s", getRelativePath(transformedAppFolder, fileToBeChanged));
             TransformationOperationException e = new TransformationOperationException(details);
             result = TOExecutionResult.error(this, e);
         }
@@ -233,77 +247,101 @@ public class InsertText extends TransformationOperation<InsertText> {
         return result;
     }
 
-    private String insertAtSpecificLine(BufferedReader readerText, BufferedReader readerOriginalFile, BufferedWriter writer) throws IOException {
+    private TOExecutionResult insertAtSpecificLine(BufferedReader readerText, BufferedReader readerOriginalFile, BufferedWriter writer, String eol) throws IOException {
         String currentLine;
-        int n = 0;
-        while((currentLine = readerOriginalFile.readLine()) != null) {
-            n++;
+        int n = 1;
+        EolBufferedReader eolReaderOriginalFile = new EolBufferedReader(readerOriginalFile);
+        boolean textInserted = false;
+
+        for (; n < lineNumber; n++) {
+            currentLine = eolReaderOriginalFile.readLineKeepEol();
+            if (currentLine == null) {
+                break;
+            }
             writer.write(currentLine);
-            writer.write(System.lineSeparator());
-            if (n == lineNumber) {
-                while((currentLine = readerText.readLine()) != null) {
-                    writer.write(currentLine);
-                    writer.write(System.lineSeparator());
-                }
+        }
+        if (n == lineNumber) {
+            textInserted = true;
+            while((currentLine = readerText.readLine()) != null) {
+                writer.write(currentLine);
+                writer.write(eol);
             }
         }
+        while((currentLine = eolReaderOriginalFile.readLineKeepEol()) != null) {
+            writer.write(currentLine);
+        }
 
-        return String.format("Text has been inserted from %s to %s at line number %d", textFileUrl, getRelativePath(), lineNumber);
+        if (textInserted) {
+            String details = String.format("Text has been inserted from %s to %s at line number %d", textFileUrl, getRelativePath(), lineNumber);
+            return TOExecutionResult.success(this, details);
+        } else {
+            String details = String.format("Text has NOT been inserted at %s because line number %d does not exist", getRelativePath(), lineNumber);
+            return TOExecutionResult.noOp(this, details);
+        }
     }
 
-    private String insertAfterRegex(BufferedReader readerText, BufferedReader readerOriginalFile, BufferedWriter writer, boolean firstOnly) throws IOException {
+    private TOExecutionResult insertAfterRegex(BufferedReader readerText, BufferedReader readerOriginalFile, BufferedWriter writer, boolean firstOnly, String eol) throws IOException {
         String currentLine;
         int n = 0;
         boolean foundFirstMatch = false;
         final Pattern pattern = Pattern.compile(regex);
-        boolean firstLine = true;
-        while((currentLine = readerOriginalFile.readLine()) != null) {
-            if(!firstLine) {
-                writer.write(System.lineSeparator());
-            }
+        EolBufferedReader eolReaderOriginalFile = new EolBufferedReader(readerOriginalFile);
+        StringBuilder readerTextStringBuilder = null;
+        String readerTextString = null;
+        if (!firstOnly) {
+            readerTextStringBuilder = new StringBuilder();
+        }
+
+        while((currentLine = eolReaderOriginalFile.readLineKeepEol()) != null) {
             writer.write(currentLine);
-            firstLine = false;
-            if((!firstOnly || !foundFirstMatch) && pattern.matcher(currentLine).matches()) {
+            if((!firstOnly || !foundFirstMatch) && pattern.matcher(removeEol(currentLine)).matches()) {
                 foundFirstMatch = true;
                 n++;
-                while((currentLine = readerText.readLine()) != null) {
-                    writer.write(System.lineSeparator());
-                    writer.write(currentLine);
-                    firstLine = false;
+                if (n == 1) {
+                    while((currentLine = readerText.readLine()) != null) {
+                        writer.write(currentLine);
+                        writer.write(eol);
+                        if (!firstOnly) {
+                            readerTextStringBuilder.append(currentLine);
+                            readerTextStringBuilder.append(eol);
+                        }
+                    }
+                    if (!firstOnly) {
+                        readerTextString = readerTextStringBuilder.toString();
+                    }
+                } else {
+                    writer.write(readerTextString);
                 }
             }
         }
 
-        String result;
-
         if (foundFirstMatch) {
-            result = String.format("Text has been inserted from %s to %s after %d line(s) that matches regular expression '%s'", textFileUrl, getRelativePath(), n, regex);
+            String details = String.format("Text has been inserted from %s to %s after %d line(s) that matches regular expression '%s'", textFileUrl, getRelativePath(), n, regex);
+            return TOExecutionResult.success(this, details);
         } else {
-            result = String.format("No text has been inserted from %s to %s, since no line has been found to match regular expression '%s'", textFileUrl, getRelativePath(), regex);
+            String details = String.format("No text has been inserted from %s to %s, since no line has been found to match regular expression '%s'", textFileUrl, getRelativePath(), regex);
+            return TOExecutionResult.noOp(this, details);
         }
-
-        return result;
     }
 
-    private String concat(BufferedReader readerText, BufferedReader readerOriginalFile, BufferedWriter writer) throws IOException {
+    private TOExecutionResult concat(BufferedReader readerText, BufferedReader readerOriginalFile, BufferedWriter writer, String eol) throws IOException {
         String currentLine;
-        boolean firstLine = true;
-        while((currentLine = readerOriginalFile.readLine()) != null) {
-            if(!firstLine) {
-                writer.write(System.lineSeparator());
-            }
+        EolBufferedReader eolReaderOriginalFile = new EolBufferedReader(readerOriginalFile);
+        boolean lastLineEndsWithEol = true;
+        while((currentLine = eolReaderOriginalFile.readLineKeepEol()) != null) {
             writer.write(currentLine);
-            firstLine = false;
+            lastLineEndsWithEol = EolHelper.endsWithEol(currentLine);
+        }
+        if (!lastLineEndsWithEol) {
+            writer.write(eol);
         }
         while((currentLine = readerText.readLine()) != null) {
-            if(!firstLine) {
-                writer.write(System.lineSeparator());
-            }
             writer.write(currentLine);
-            firstLine = false;
+            writer.write(eol);
         }
 
-        return String.format("Text has been inserted from %s to %s at the end of the file", textFileUrl, getRelativePath());
+        String details = String.format("Text has been inserted from %s to %s at the end of the file", textFileUrl, getRelativePath());
+        return TOExecutionResult.success(this, details);
     }
 
     @Override
